Added system for describing State Machine Elements and a UI to display the info to users.  Cleaned up a lot of things in the UI, extraneous Propertes, Elements that didn't do anything, etc.  Added Variables which should be used for internal logic while parameters are used for getting input from outside.  Added capability to transition directly to a result from the start bar in a StateMachine.  Changed newClick such that it doesn't update anything internally, but the operator now needs to set newClick back to false before checking it again.  Discovered an execution order bug that will force us to be pretty creative when analyzing sessions run before the current version (syntax 0.0.1, picto 0.0.12).  Init Transitions were being triggered before the EntryScript on their level was called.  This may make it impossible to run AnalysisEntryScripts at the right time.  Fixed this such that in future versions, EntryScript is always called before InitTransitions are sent to the server.

+++ source/common/parameter/NumericParameter.h
@@ -45,7 +45,7 @@ public:
 	int getValue(){return propertyContainer_->getPropertyValue("Value").toInt();};
 	void setValue(int val){propertyContainer_->setPropertyValue("Value",val);};
 
-	virtual QString friendlyTypeName(){return "Numeric";};
+	virtual QString friendlyTypeName(){return "Integer";};
 	//note that the lessThan & greaterThan functions aren't redefined, 
 	//so they will always return false
 	//bool greaterThan(Parameter& RHS);
